feat(components): create TodoList and TodoForm components and restructure app
- Move todos array from App to TodoList

- Extract list rendering logic to TodoList component

- Add TodoList component to application structure

- Update imports in App.jsx
