+++ src/App.jsx
@@ -1,22 +1,16 @@
-import './App.css'
+import './App.css';
+import TodoList from './TodoList.jsx';
+import TodoForm from './TodoForm.jsx';
 
-/*extract from app.js*/
 function App() {
-const todos = [
-  {id: 1, title: "review resources"},
-  {id: 2, title: "take notes"},
-  {id: 3, title: "code out app"}
-]
-
 
   return (
    <div>
     <h1>My Todos</h1>
-    <ul>
-      {todos.map(todo => <li key={todo.id}>{todo.title}</li>)}
-    </ul>
+    <TodoList />
+    <TodoForm />
    </div>
-  )
-};
+  );
+}
 
-export default App
+export default App;
