+++ src/TodoForm.jsx
@@ -1,0 +1,12 @@
+function TodoForm() {
+        return (
+            <form> 
+                <label htmlFor="todoTitle">Todo</label>
+                <input id="todoTitle" type="text" />
+                <button type ="submit">Add Todo</button>
+            </form>
+        );
+    }
+
+
+export default TodoForm;
